Acceleration/deceleration now 32 bit wide
Fix in original repetier firmware
https://github.com/repetier/Repetier-Firmware/commit/3ae42d2d0960e5d965bc73f048040300d0e65e1a
nun auch hier.

+++ RF1000/Repetier/motion.h
@@ -72,8 +72,8 @@ private:
     float				minSpeed;
     float				distance;
     ticks_t				fullInterval;				///< interval at full speed in ticks/step.
-    uint16_t			accelSteps;					///< How much steps does it take, to reach the plateau.
-    uint16_t			decelSteps;					///< How much steps does it take, to reach the end speed.
+    uint32_t 			accelSteps;					///< How much steps does it take, to reach the plateau.
+    uint32_t 			decelSteps;					///< How much steps does it take, to reach the end speed.
     uint32_t			accelerationPrim;			///< Acceleration along primary axis
     uint32_t			fAcceleration;				///< accelerationPrim*262144/F_CPU
     speed_t				vMax;						///< Maximum reached speed in steps/s.
@@ -417,7 +417,7 @@ public:
 
     inline bool moveDecelerating()
     {
-        if(stepsRemaining <= decelSteps)
+        if(stepsRemaining <= static_cast<int32_t>(decelSteps))
         {
             if (!(flags & FLAG_DECELERATING))
             {
